+++ .idea/inspectionProfiles/Project_Default.xml
@@ -1,0 +1,18 @@
+<component name="InspectionProjectProfileManager">
+  <profile version="1.0">
+    <option name="myName" value="Project Default" />
+    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
+      <option name="ignoredPackages">
+        <list>
+          <option value="fdb" />
+          <option value="pyinstaller" />
+          <option value="APScheduler" />
+          <option value="SQLAlchemy" />
+          <option value="python-dotenv" />
+          <option value="pandas" />
+          <option value="requests" />
+        </list>
+      </option>
+    </inspection_tool>
+  </profile>
+</component>
