+++ core/collection/admin.py
