+++ core/collection/views.py
