+++ core/messaging/consumer.py
