--- infra/__init__.py
@@ -1,1 +1,0 @@
-# Infra module
